Revert safe and working shell escaping in sendmail delivery
  Commit 81e9940b0e7a8140ea8178ced03c071c91f756e9 introduced a broken
and less safe escaping mechanism for sendmail command arguments, this
change revert a version more similar as it was just before. Arguments
may be quoted or escaped, but not both.
  We remove example verifying that destinations are quoted, but many
other examples need to be changed back to prior expectations (without
quotes), as they were changed by 81e9940b to force those other tests to
pass.